Fixed compiler warning in test host.

+++ Test Host/ViewController.m
@@ -46,7 +46,7 @@
         
         NSHTTPURLResponse *HTTPResponse = (NSHTTPURLResponse *)response;
         if (HTTPResponse.statusCode != 200) {
-            [[[UIAlertView alloc] initWithTitle:[NSString stringWithFormat:@"%d status code", HTTPResponse.statusCode] message:nil delegate:nil cancelButtonTitle:@"Close" otherButtonTitles:nil] show];
+            [[[UIAlertView alloc] initWithTitle:[NSString stringWithFormat:@"%ld status code", (long)HTTPResponse.statusCode] message:nil delegate:nil cancelButtonTitle:@"Close" otherButtonTitles:nil] show];
             return;
         }
